requirements.txt
certifi==2019.9.11
chardet==3.0.4
idna==2.8
requests==2.22.0
RPi.GPIO==0.7.0
tuyapy==0.1.3
tuyasmartapi==0.1.2
urllib3==1.25.5

